backend/requirements.txt
flask
flask-cors
pandas
scikit-learn
joblib
prometheus-client
gunicorn
pyjwt